TCA: Remove force-inlining of functions whose body is not always available
Beginning with gcc 4.7 force-inlining of functions that don't have the
function body available in some situations the functions are used result in
an error.

gcc 4.6 silently ignored the inline keyword for functions that could not
be inlined.

This commit resolves the compile errors of the type:
error: inlining failed in call to always_inline 'get_extdsp_global': function body not available